API/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
httpx==0.25.0
pydantic==1.10.7
python-multipart==0.0.6
passlib==1.7.4
